requirements.txt
azure-identity>=1.12.0
azure-mgmt-datafactory>=2.0.0
requests>=2.28.0 